refactor(dashmate): move the release packaging scripts next to the package
The deb versioning scripts lived at the repository root under snake_case
names while the package they serve is dashmate, and the release workflow
carried a hundred lines of inline bash that could only be exercised by
pushing a tag.

The three scripts move to packages/dashmate/scripts/ under kebab-case
names, matching the scripts already there. The dashmate package is ESM,
so the two JS files trade module.exports and require.main for exports and
an import.meta.url main check; nothing else about them changes.

Three workflow steps become invocations of scripts a release engineer can
run locally: the version gate, the check that the built deb carries the
validated version, and the checksum recording. Every comment, error
message and exit code moves with the logic it explains.

The checksums job's three remaining bash steps stay inline. That job
deliberately has no checkout so that no repository script runs beside the
signing material, and extracting them would mean adding one.

release.yml: 762 -> 662 lines.
