+++ packages/dashmate/test/unit/packaging/debVersion.spec.js
@@ -1,8 +1,8 @@
 import { execFileSync } from 'node:child_process';
 import { fileURLToPath } from 'node:url';
-import { debVersionFromSemver, debFileNameVersion } from '../../../../../scripts/deb_version.js';
+import { debVersionFromSemver, debFileNameVersion } from '../../../scripts/deb-version.js';
 
-const SCRIPT_PATH = fileURLToPath(new URL('../../../../../scripts/deb_version.js', import.meta.url));
+const SCRIPT_PATH = fileURLToPath(new URL('../../../scripts/deb-version.js', import.meta.url));
 
 /**
  * Independent port of dpkg's version comparison (`verrevcmp` in dpkg's version.c),
@@ -112,7 +112,7 @@ function compareDebVersions(left, right) {
   return upstream === 0 ? verrevcmp(a.revision, b.revision) : upstream;
 }
 
-describe('deb_version.js', () => {
+describe('deb-version.js', () => {
   describe('version comparison oracle', () => {
     it('should order versions the way dpkg does', () => {
       expect(compareDebVersions('4.1.0', '4.1.0')).to.equal(0);
